+++ crates/criticalup-core/src/download_server_client.rs
@@ -52,6 +52,7 @@ impl DownloadServerClient {
         .await
     }
 
+    #[tracing::instrument(level = "debug", skip_all)]
     pub async fn get_keys(&self) -> Result<Keychain, Error> {
         let resp: KeysManifest = self
             .json(self.send(self.client.get(self.url("/v1/keys"))).await?)
@@ -61,6 +62,10 @@ impl DownloadServerClient {
         Ok(keychain)
     }
 
+    #[tracing::instrument(level = "debug", skip_all, fields(
+        %product,
+        %release,
+    ))]
     pub async fn get_product_release_manifest(
         &self,
         product: &str,
@@ -74,6 +79,12 @@ impl DownloadServerClient {
         .await
     }
 
+    #[tracing::instrument(level = "debug", skip_all, fields(
+        %product,
+        %release,
+        %package,
+        %format
+    ))]
     pub async fn download_package(
         &self,
         product: &str,
@@ -86,6 +97,7 @@ impl DownloadServerClient {
         let download_url =
             format!("/v1/releases/{product}/{release}/download/{package}/{artifact_format}");
 
+        tracing::info!("Downloading component '{package}' for '{product}' ({release})",);
         let response = self
             .send_with_auth(self.client.get(self.url(download_url.as_str())))
             .await?;